WIP: Clean code and remove unnecessary resources
