+++ app/src/androidTest/java/com/anibalbastias/android/vuro/dualcamerapp/MainActivityTest.kt
@@ -7,7 +7,7 @@ import androidx.test.espresso.matcher.ViewMatchers.withId
 import androidx.test.ext.junit.rules.ActivityScenarioRule
 import androidx.test.ext.junit.runners.AndroidJUnit4
 import com.anibalbastias.android.vuro.R
-import com.anibalbastias.android.vuro.dualcamerapp.split.MainActivity
+import com.anibalbastias.android.vuro.dualcamerapp.ui.MainActivity
 import org.junit.Rule
 import org.junit.Test
 import org.junit.runner.RunWith
@@ -19,7 +19,8 @@ import org.junit.runner.RunWith
 class MainActivityTest {
 
     @get:Rule
-    val rule = ActivityScenarioRule<MainActivity>(MainActivity::class.java)
+    val rule = ActivityScenarioRule<MainActivity>(
+        MainActivity::class.java)
 
     @Test
     fun launchMainActivity() {
