+++ app/src/main/res/drawable/splashscreen.xml
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="utf-8"?>
 <layer-list xmlns:android="http://schemas.android.com/apk/res/android">
-    <item android:drawable="@color/colorPrimary" />
+    <item android:drawable="@color/primaryColor" />
     <item
         android:bottom="0dp"
         android:left="340dp"
